Add files via upload
This is my starting efforts working with the OpenBiomechanics created by Driveline Baseball (I do not own any of the data). The plot I have created using python was made to examine the effects of max torso rotational velocity on ball speed from the pitcher. From the findings, there is a strong correlation between an increase in torso rotational velocity and pitch velocity shown in the graph.